Add comprehensive face recognition tests (real + mocked)
## Real Integration Tests (NEW)
- Test actual TFLite model loading and inference
- Test real OpenCV face detection
- End-to-end tests with synthetic face images
- Configuration validation tests

## Mock Unit Tests (NEW)
- Fast tests for CI/CD
- Test signal auto-generation
- Test error handling (no face, multiple faces, corrupted images)
- Test lazy loading behavior

## Bug Fixes
- Fix FLOAT64 -> FLOAT32 dtype issue in preprocessing
- Ensure TFLite model receives correct data types

## Test Infrastructure
- Created tests/fixtures/test_images/ for test photos
- Auto-generate synthetic face images for testing
- README for adding real face photos
- Updated .gitignore to protect privacy

## Test Results
- 10/11 integration tests PASS
- Model loads correctly
- Embeddings generate with proper L2 normalization
- Signal auto-generates embeddings

## Usage
Fast tests: pytest tests/unit/test_face_recognition.py
Real tests: pytest tests/integration/test_face_recognition_real.py -v
All tests: pytest tests/

## Next Steps
Add real human face photo to tests/fixtures/test_images/real_face_frontal.jpg
for more realistic testing

+++ ml_models/face_recognition/inference/mobilefacenet.py
@@ -59,12 +59,12 @@ class MobileFaceNet(BaseFaceRecognitionModel):
         image = cv2.resize(image, (112, 112))
 
         # Normalize: (pixel - 127.5) / 128.0
-        mean = np.array(MOBILEFACENET_CONFIG["mean"])
-        std = np.array(MOBILEFACENET_CONFIG["std"])
+        mean = np.array(MOBILEFACENET_CONFIG["mean"], dtype=np.float32)
+        std = np.array(MOBILEFACENET_CONFIG["std"], dtype=np.float32)
         image = (image.astype(np.float32) - mean) / std
 
-        # Add batch dimension
-        image = np.expand_dims(image, axis=0)
+        # Add batch dimension and ensure float32
+        image = np.expand_dims(image, axis=0).astype(np.float32)
 
         return image
